+++ src/app/ui/task-listitem.tsx
@@ -1,5 +1,5 @@
 import { EditIcon, TrashIcon } from "@/app/ui/svg-icons";
-import { Todo } from "@/app/models/todoItem";
+import { Todo } from "@/app/lib/models/todoItem";
 import Link from "next/link";
 
 type TaskListItemProps = {
